--- src/Components/Frontoffice/Testimonials.js
@@ -1,25 +1,0 @@
-import Card from '../Cards/Card/Card'
-import { Get } from '../../Services/publicApiService'
-import { useState, useEffect } from 'react'
-
-const Testimonials = () => {
-    const endpoint = process.env.REACT_APP_BASE_URL + process.env.REACT_APP_TESTIMONIALS
-    const [testimonials, setTestimonials] = useState([])
-
-    const fetchTestimonials = async () => {
-        const res = await Get(endpoint)
-        console.log(res)
-    }
-
-    useEffect(() => {
-        fetchTestimonials()
-    }, [])
-
-    return(
-        <div>
-
-        </div>
-    )
-}
-
-export default Testimonials
