+++ src/Components/Frontoffice/Testimonials/Testimonials.css
@@ -1,0 +1,19 @@
+.testimonials-container {
+    display: flex;
+    flex-direction: row;
+    flex-wrap: wrap;
+    margin: auto;
+}
+
+.testimonial-card {
+    display: flex;
+    width: 20%;
+    margin: 15px
+}
+
+.testimonials-title {
+    font-family: 'Poppins', sans-serif;
+    font-size: 3.5rem;
+    text-align: center;
+    margin-top: 10px;
+}
